Spring and Hibernate configuration added
