+++ src/main/java/com/extra/email/SendMailSSL.java
@@ -1,0 +1,48 @@
+package com.extra.email;
+
+import java.util.Properties;
+
+import javax.mail.Message;
+import javax.mail.MessagingException;
+import javax.mail.PasswordAuthentication;
+import javax.mail.Session;
+import javax.mail.Transport;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeMessage;
+
+public class SendMailSSL {
+	public static void main(String[] args) {
+		Properties props = new Properties();
+		props.put("mail.smtp.host", "smtp.gmail.com");
+		props.put("mail.smtp.socketFactory.port", "465");
+		props.put("mail.smtp.socketFactory.class",
+				"javax.net.ssl.SSLSocketFactory");
+		props.put("mail.smtp.auth", "true");
+		props.put("mail.smtp.port", "465");
+ 
+		Session session = Session.getDefaultInstance(props,
+			new javax.mail.Authenticator() {
+				protected PasswordAuthentication getPasswordAuthentication() {
+					return new PasswordAuthentication("muku3011","lalita_3011");
+				}
+			});
+ 
+		try {
+ 
+			Message message = new MimeMessage(session);
+			message.setFrom(new InternetAddress("muku3011@gmail.com"));
+			message.setRecipients(Message.RecipientType.TO,
+					InternetAddress.parse("mukesh.bciit@gmail.com"));
+			message.setSubject("Testing Subject");
+			message.setText("Dear Mail Crawler," +
+					"\n\n No spam to my email, please!");
+ 
+			Transport.send(message);
+ 
+			System.out.println("Done");
+ 
+		} catch (MessagingException e) {
+			throw new RuntimeException(e);
+		}
+	}
+}
