+++ src/main/java/com/extra/email/mail_sms/UserDetail.java
@@ -1,0 +1,20 @@
+package com.extra.email.mail_sms;
+
+public class UserDetail
+{
+	private String number;
+	private String mailId;
+	
+	public String getNumber() {
+		return number;
+	}
+	public void setNumber(String number) {
+		this.number = number;
+	}
+	public String getMailId() {
+		return mailId;
+	}
+	public void setMailId(String mailId) {
+		this.mailId = mailId;
+	}
+}
