+++ src/main/java/com/extra/quartz_job/listener/HelloJob.java
@@ -1,0 +1,16 @@
+package com.extra.quartz_job.listener;
+
+import org.quartz.Job;
+import org.quartz.JobExecutionContext;
+import org.quartz.JobExecutionException;
+
+public class HelloJob implements Job
+{
+	public void execute(JobExecutionContext context)
+	throws JobExecutionException {
+ 
+		System.out.println("Hello Quartz!");	
+ 
+	}
+ 
+}
