+++ src/main/java/com/extra/spring/FirstBean.java
@@ -1,0 +1,14 @@
+package com.extra.spring;
+
+public class FirstBean {
+
+    private String message;
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public void getMessage() {
+        System.out.println("Your Message : " + message);
+    }
+}
