+++ src/main/java/com/ocp/topics/design_pattern/Singleton.java
@@ -1,0 +1,4 @@
+package com.ocp.topics.design_pattern;
+
+public class Singleton {
+}
