+++ src/main/java/com/ocp/topics/jdbc/JdbcConnection.java
@@ -5,14 +5,15 @@ import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.Statement;
 
-class OracleCon {
+public class JdbcConnection {
 	public static void main(String args[]) {
 		try {
 			// step1 load the driver class
 			Class.forName("oracle.jdbc.driver.OracleDriver");
 
 			// step2 create the connection object
-			Connection con = DriverManager.getConnection("jdbc:oracle:thin:@203.122.59.147:1521:DP9A", "dpownera","dpownera");
+			Connection con = DriverManager.getConnection("jdbc:oracle:thin:@127.0.0.1:1521:xe", "testdb","testdb");
+
 
 			// step3 create the statement object
 			Statement stmt = con.createStatement();
@@ -26,7 +27,7 @@ class OracleCon {
 			con.close();
 
 		} catch (Exception e) {
-			System.out.println(e);
+			System.out.println(e.getLocalizedMessage());
 		}
 
 	}
