+++ src/main/java/com/ocp/topics/serialization/Externalization.java
@@ -1,0 +1,4 @@
+package com.ocp.topics.serialization;
+
+public class Externalization {
+}
