+++ src/main/java/com/ocp/topics/serialization/Serialization.java
@@ -1,0 +1,7 @@
+package com.ocp.topics.serialization;
+
+/**
+ * Created by joshimuk on 8/3/2016.
+ */
+public class Serialization {
+}
