--- src/main/java/com/others/programs/FibonaciSeries.java
@@ -1,4 +1,0 @@
-package com.others.programs;
-
-public class FibonaciSeries {
-}
