--- src/main/java/com/others/quartz_job/listener/HelloJob.java
@@ -1,16 +1,0 @@
-package com.others.quartz_job.listener;
-
-import org.quartz.Job;
-import org.quartz.JobExecutionContext;
-import org.quartz.JobExecutionException;
-
-public class HelloJob implements Job
-{
-	public void execute(JobExecutionContext context)
-	throws JobExecutionException {
- 
-		System.out.println("Hello Quartz!");	
- 
-	}
- 
-}
