+++ src/main/resources/notes/Array Notes
