+++ src/main/resources/notes/Variables Notes
